monkey_island/requirements.txt
python-dateutil
tornado
werkzeug
jinja2
markupsafe
itsdangerous
click
flask
Flask-Pymongo
Flask-Restful
jsonschema
netifaces
ipaddress
enum34
PyCrypto